feat: improve repo organization and responsiveness
