--- .env.local.example
@@ -1,9 +1,0 @@
-# SPDX-FileCopyrightText: 2024 PNED G.I.E.
-#
-# SPDX-License-Identifier: Apache-2.0
-NEXTAUTH_URL=http://localhost:3000
-NEXTAUTH_SECRET=your-secret
-KEYCLOAK_CLIENT_ID=ckan
-KEYCLOAK_ISSUER_URL=https://keycloak-test.healthdata.nl/realms/ckan
-END_SESSION_URL=https://keycloak-test.healthdata.nl/realms/ckan/protocol/openid-connect/logout
-REFRESH_TOKEN_URL=https://keycloak-test.healthdata.nl/realms/ckan/protocol/openid-connect/token
