+++ package.json
@@ -26,7 +26,6 @@
     "clsx": "^2.1.0",
     "cryptr": "^6.3.0",
     "jwt-decode": "^4.0.0",
-    "lucide-react": "^0.331.0",
     "next": "14.1.0",
     "next-auth": "^4.24.5",
     "react": "^18",
