+++ src/app/api/auth/logout/route.ts
@@ -1,9 +1,8 @@
 // SPDX-FileCopyrightText: 2024 PNED G.I.E.
 //
 // SPDX-License-Identifier: Apache-2.0
-import { getToken } from '@/utils/auth';
+import { authOptions, getToken } from '@/utils/auth';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../[...nextauth]/route';
 
 export async function GET() {
   const session = await getServerSession(authOptions);
