+++ src/utils/auth.ts
@@ -1,17 +1,68 @@
 // SPDX-FileCopyrightText: 2024 PNED G.I.E.
 //
 // SPDX-License-Identifier: Apache-2.0
+import { encrypt } from '@/utils/encryption';
 import { jwtDecode } from 'jwt-decode';
-import type { Session } from 'next-auth';
+import type { NextAuthOptions, Session } from 'next-auth';
 import { Account, getServerSession } from 'next-auth';
 import type { JWT } from 'next-auth/jwt';
+import Keycloack from 'next-auth/providers/keycloak';
 import { signOut } from 'next-auth/react';
-import { authOptions } from '../app/api/auth/[...nextauth]/route';
 import { decrypt } from './encryption';
 
 type ExtendedSession = Session & { id_token: string; access_token: string; error?: string };
 
-async function getToken(tokenType: 'access_token' | 'id_token') {
+type JWTCallbackEntry = {
+  token: JWT;
+  account: Account | null;
+};
+
+type SessionCallbackEntry = {
+  token: JWT;
+  session: Session;
+};
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    Keycloack({
+      clientId: `${process.env.KEYCLOAK_CLIENT_ID}`,
+      clientSecret: `${process.env.KEYCLOAK_CLIENT_SECRET}`,
+      issuer: process.env.KEYCLOAK_ISSUER_URL,
+      authorization: { params: { scope: 'openid profile email offline_access' } },
+    }),
+  ],
+  callbacks: {
+    async jwt({ token, account }: JWTCallbackEntry) {
+      const currTimestamp = Math.floor(Date.now() / 1000);
+      const isTokenExpired = token?.expires_at && (token?.expires_at as number) <= currTimestamp;
+
+      if (account) {
+        return completeTokenWithAccountInfo(token, account);
+      } else if (!isTokenExpired) {
+        return token;
+      } else {
+        try {
+          const refreshedToken = await refreshAccessToken(token);
+          return refreshedToken;
+        } catch (error) {
+          return { ...token, error: 'RefreshAccessTokenError' };
+        }
+      }
+    },
+
+    async session({ session, token }: SessionCallbackEntry) {
+      return {
+        ...session,
+        access_token: encrypt(token.access_token as string),
+        id_token: encrypt(token.id_token as string),
+        roles: (token.decoded as { realm_access?: { roles?: string[] } }).realm_access?.roles,
+        error: token.error,
+      };
+    },
+  },
+};
+
+export async function getToken(tokenType: 'access_token' | 'id_token') {
   const session = (await getServerSession(authOptions)) as ExtendedSession;
   if (session) {
     const tokenDecrypted = decrypt(session[tokenType]!);
@@ -20,7 +71,7 @@ async function getToken(tokenType: 'access_token' | 'id_token') {
   return null;
 }
 
-function completeTokenWithAccountInfo(token: JWT, account: Account): JWT {
+export function completeTokenWithAccountInfo(token: JWT, account: Account): JWT {
   return {
     ...token,
     decoded: jwtDecode(account.access_token!) as string,
@@ -31,7 +82,7 @@ function completeTokenWithAccountInfo(token: JWT, account: Account): JWT {
   };
 }
 
-async function refreshAccessToken(token: JWT) {
+export async function refreshAccessToken(token: JWT) {
   const response = await fetch(`${process.env.REFRESH_TOKEN_URL}`, {
     headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
     body: new URLSearchParams({
@@ -54,7 +105,7 @@ async function refreshAccessToken(token: JWT) {
   };
 }
 
-async function keycloackSessionLogOut() {
+export async function keycloackSessionLogOut() {
   try {
     await fetch('/api/auth/logout');
   } catch (error) {
@@ -62,10 +113,8 @@ async function keycloackSessionLogOut() {
   }
 }
 
-function logOutIfSessionError(session: ExtendedSession | null, status: string) {
+export function logOutIfSessionError(session: ExtendedSession | null, status: string) {
   if (session && status !== 'loading' && session?.error) {
     signOut({ callbackUrl: '/' });
   }
 }
-
-export { completeTokenWithAccountInfo, getToken, keycloackSessionLogOut, logOutIfSessionError, refreshAccessToken };
